requirements.txt
base58==2.1.1
fastecdsa==2.2.3
Flask==2.0.2
pycryptodome==3.12.0
qrcode==7.3.1
